requirements/dev.txt
-r base.txt
pytest==9.0.2
pytest-django==4.12.0
factory-boy==3.3.3
black==25.1.0
isort==6.0.1
flake8==7.2.0
pre-commit==4.2.0
coverage==7.6.0
